Proof of concept for a JsonCell type for Bobtail.
Essentially the idea is to have a cell type that mimics
regular JSON objects. No need to call .get() or .set();
ES2015's Proxies allow regular object setting and getting
to trigger events and record dependencies.

Currently, binding and refreshing are very inefficient.
e.g., bind -> w.x.y.z will check to see if it needs to
refresh on any changes to any of the keys in that
change--though of course it usually will not.

Finally, the API is still very much in flux. Numerous
support methods and functions, e.g. for casting, still need
to be added.