Breaking change.
Rename JsonCell to ObsJsonCell. Add Dep and SrcJsonCell classes,
analagous to other Dep/Src classes in bobtail. DepJsonCell is read-
only; attempting to write to or edit any fields will lead to an ex-
ception.

In addition, we now support ObsJsonCells with primitive values. That
is, an ObsJsonCell can now have a data field with a String, Number,
Symbol, null, undefined, or boolean value.