Return proxies on the fly. Add raw method.
This commit does a few things. First, the code has been
restructured so that rather than storing the object as a proxy,
the data accessor creates a new proxy and returns it every time.
This is because Proxies are cheap, and it allows us to maintain
our object's state internally without the Proxy. It also means
that we can return safe, non-reactive copies of or references to
the underlying data, which was previously impossible.

The previous version had a subtle breakage where we relied on
this._base, believing that it would provide a non-proxied
representation of the underlying object, when in fact, if assigned
to null, it would no longer do so. This led to the bug the
nullification test case guards against. The Proxy's has method
references this._base, which remains null, rather than the array,
which causes the built in Array.map method to fail.

In addition, we add a cloneRaw() method which returns a non-reactive
clone of the object.